change ChainIcons parent dir
