+++ src/components/Icons/Svg/ChainIcons/Bsc.tsx
@@ -2,10 +2,10 @@
 // @ts-ignore
 import React from 'react';
 
-import { SVGProps } from '../svgInterface';
+import { SVGProps } from '../FT/svgInterface';
 
 /* eslint-disable max-len */
-export function BscBridged(props: SVGProps) {
+export function Bsc(props: SVGProps) {
   const { size, ...rest } = props;
   return (
     <svg
@@ -21,52 +21,52 @@ export function BscBridged(props: SVGProps) {
         cy="6"
         r="6"
         transform="matrix(-1 0 0 1 12 0)"
-        fill="#FF0000"
+        fill="#1C1C1C"
       />
-      <g clipPath="url(#clip0_6150_47860)">
+      <g clip-path="url(#clip0_6150_47876)">
         <path
           d="M5.42814 6.92349C5.42814 6.74863 5.42703 6.57378 5.42898 6.39892C5.42944 6.35641 5.41537 6.33191 5.37815 6.31028C5.08376 6.13914 4.79066 5.96586 4.49719 5.79323C4.48728 5.78738 4.47784 5.78079 4.46756 5.77587C4.44358 5.76455 4.43488 5.74543 4.43673 5.72C4.4372 5.71434 4.43673 5.70859 4.43673 5.70283C4.43673 5.51359 4.43803 5.32444 4.43571 5.13529C4.43516 5.09204 4.45025 5.06818 4.487 5.04702C4.64948 4.95356 4.81075 4.85806 4.97118 4.76126C5.00266 4.74232 5.02617 4.74371 5.057 4.76218C5.35279 4.93945 5.64931 5.11524 5.94574 5.2914C5.95556 5.29724 5.96565 5.30263 5.97528 5.30885C5.99009 5.31831 6.00435 5.31859 6.01944 5.30922C6.03517 5.29938 6.05156 5.29047 6.06758 5.281C6.35827 5.10921 6.64924 4.93797 6.9391 4.76478C6.97456 4.74362 7.00085 4.74334 7.03621 4.76478C7.19526 4.86112 7.3556 4.95551 7.51669 5.04842C7.54965 5.06744 7.5615 5.09 7.5615 5.12731C7.5615 5.31794 7.5627 5.50858 7.56631 5.69912C7.56714 5.74191 7.55335 5.76622 7.51613 5.78785C7.23044 5.95389 6.94595 6.12206 6.66109 6.28968C6.64257 6.30054 6.62461 6.31251 6.60554 6.32216C6.58221 6.33404 6.57286 6.35242 6.57397 6.37785C6.57471 6.39502 6.57388 6.41228 6.57388 6.42945C6.57388 6.76627 6.57388 7.10308 6.57379 7.43999C6.57379 7.51721 6.57703 7.50635 6.51371 7.54366C6.35845 7.63526 6.20237 7.72538 6.04795 7.81838C6.01388 7.83889 5.98787 7.83824 5.95417 7.81801C5.79475 7.72232 5.63431 7.62812 5.47314 7.5354C5.44036 7.51656 5.4274 7.49428 5.42777 7.4566C5.42925 7.27887 5.42842 7.10113 5.42842 6.9234H5.42833L5.42814 6.92349Z"
-          fill="white"
+          fill="#F0B90B"
         />
         <path
           d="M8.54875 5.6709C8.55023 5.69475 8.55236 5.71396 8.55236 5.73327C8.55255 6.15315 8.55255 6.57312 8.55255 6.993C8.55255 7.14633 8.55144 7.29974 8.55329 7.45307C8.55375 7.49205 8.54135 7.51525 8.507 7.53539C8.0181 7.82144 7.52994 8.10888 7.04159 8.39585C7.02854 8.40346 7.01474 8.40977 6.99512 8.4198C6.99308 8.3991 6.99067 8.38555 6.99067 8.37191C6.9904 8.17988 6.99123 7.98785 6.98993 7.79582C6.98966 7.75981 7.00299 7.73856 7.03372 7.72055C7.33117 7.54635 7.62751 7.3701 7.92534 7.19654C7.96385 7.17408 7.97829 7.14865 7.9782 7.10465C7.97736 6.76933 7.97875 6.434 7.97949 6.09857C7.97949 6.07704 7.9807 6.05551 7.97959 6.03407C7.97838 6.00948 7.98958 5.99435 8.01014 5.98228C8.1789 5.88269 8.3473 5.78246 8.51607 5.68278C8.52403 5.67804 8.53375 5.67628 8.54894 5.67099L8.54875 5.6709Z"
-          fill="white"
+          fill="#F0B90B"
         />
         <path
           d="M3.45884 5.67188C3.47236 5.67874 3.48624 5.68478 3.49921 5.69248C3.65807 5.7864 3.81628 5.88135 3.97589 5.97398C4.00857 5.99291 4.02097 6.01528 4.02097 6.05296C4.02051 6.3998 4.02143 6.74663 4.02245 7.09357C4.02264 7.17273 4.01551 7.16104 4.08318 7.20095C4.37508 7.37339 4.66679 7.54621 4.95971 7.71707C4.99655 7.73861 5.01099 7.76255 5.01053 7.80552C5.00831 7.9918 5.00961 8.17816 5.00951 8.36453C5.00951 8.38003 5.00822 8.39553 5.0072 8.4192C4.98739 8.4101 4.9722 8.40462 4.9585 8.39664C4.77955 8.29176 4.60078 8.18652 4.42192 8.08145C4.11605 7.90158 3.81027 7.72181 3.50448 7.54194C3.49587 7.53683 3.48745 7.53136 3.47856 7.52681C3.45736 7.51595 3.44634 7.49971 3.44801 7.47511C3.44884 7.46231 3.44801 7.44931 3.44801 7.43641C3.44801 6.87026 3.44801 6.30411 3.44801 5.73796C3.44801 5.71819 3.44801 5.69842 3.44801 5.67865C3.45171 5.67642 3.45542 5.6742 3.45912 5.67197L3.45884 5.67188Z"
-          fill="white"
+          fill="#F0B90B"
         />
         <path
           d="M7.55456 3.91293C7.53744 3.92453 7.52725 3.93223 7.51642 3.93863C7.35404 4.03507 7.1911 4.13048 7.02956 4.2283C6.9978 4.24751 6.97429 4.24501 6.94374 4.22681C6.64629 4.04973 6.348 3.87385 6.0499 3.6977C6.04009 3.69185 6.02991 3.68647 6.02028 3.68034C6.00547 3.67087 5.99121 3.6705 5.97612 3.67997C5.96158 3.68906 5.94649 3.69742 5.93177 3.70614C5.64006 3.87905 5.34807 4.05131 5.0572 4.22561C5.02285 4.24621 4.99767 4.24593 4.96388 4.22542C4.80613 4.12964 4.6468 4.03646 4.48812 3.94216C4.47507 3.93437 4.46285 3.92518 4.44461 3.91283C4.46118 3.90123 4.47183 3.89251 4.48359 3.88555C4.97619 3.59421 5.46898 3.30325 5.96121 3.01145C5.98843 2.9953 6.00917 2.9965 6.03574 3.01228C6.52945 3.30445 7.02372 3.5957 7.51772 3.8874C7.52846 3.89371 7.53827 3.9016 7.55438 3.91293H7.55456Z"
-          fill="white"
+          fill="#F0B90B"
         />
         <path
           d="M7.98144 5.5037C7.97968 5.47929 7.97737 5.46128 7.97728 5.44337C7.977 5.26861 7.97709 5.09384 7.97709 4.91908C7.97709 4.90618 7.97598 4.89318 7.97709 4.88037C7.98005 4.84789 7.96617 4.82794 7.93784 4.81151C7.79573 4.72881 7.65464 4.64435 7.51319 4.56036C7.48736 4.54505 7.46143 4.52982 7.43598 4.51395C7.42903 4.50968 7.42357 4.503 7.41376 4.49409C7.42542 4.48472 7.43422 4.4759 7.44449 4.46978C7.61039 4.37102 7.77666 4.27301 7.94228 4.17389C7.96644 4.15941 7.98598 4.1609 8.00949 4.17491C8.179 4.27589 8.34897 4.37613 8.51932 4.47572C8.54339 4.48982 8.55246 4.50764 8.55237 4.53493C8.55172 4.73401 8.55172 4.93318 8.55237 5.13236C8.55237 5.15955 8.54403 5.17756 8.51978 5.19176C8.35064 5.2906 8.18215 5.39056 8.01338 5.48996C8.00616 5.49423 7.99792 5.49664 7.98135 5.5037H7.98144Z"
-          fill="white"
+          fill="#F0B90B"
         />
         <path
           d="M4.01477 5.50587C3.9958 5.49613 3.98311 5.49047 3.97126 5.48341C3.8112 5.38865 3.65159 5.29315 3.49097 5.19941C3.46033 5.18168 3.44653 5.16108 3.44681 5.1247C3.44829 4.93118 3.4482 4.73767 3.44681 4.54416C3.44653 4.50824 3.45885 4.48671 3.48986 4.4687C3.65428 4.37292 3.81786 4.27575 3.98089 4.17755C4.00903 4.16057 4.03125 4.16029 4.05958 4.17727C4.22409 4.27575 4.38953 4.37264 4.5545 4.47019C4.56413 4.47585 4.57265 4.48337 4.58737 4.49413C4.57089 4.50601 4.5583 4.5165 4.54441 4.52476C4.38805 4.61739 4.23242 4.71122 4.07439 4.80087C4.03209 4.82482 4.01848 4.85285 4.01913 4.89991C4.02153 5.08191 4.02024 5.26401 4.01987 5.44601C4.01987 5.46374 4.0169 5.48146 4.01468 5.50587H4.01477Z"
-          fill="white"
+          fill="#F0B90B"
         />
         <path
           d="M5.42733 7.9978C5.45121 8.0107 5.46612 8.01813 5.48047 8.02667C5.63794 8.11985 5.79606 8.2121 5.95224 8.30742C5.98677 8.32849 6.01214 8.32636 6.04537 8.30631C6.20211 8.21183 6.36014 8.11957 6.51779 8.02676C6.53205 8.01841 6.54705 8.01126 6.56843 8.00003C6.57056 8.0211 6.57306 8.03456 6.57316 8.04801C6.57343 8.23995 6.5726 8.43188 6.5739 8.62382C6.57417 8.65853 6.56094 8.67839 6.53168 8.69538C6.36606 8.79162 6.201 8.88889 6.03676 8.98745C6.00816 9.00462 5.98668 9.00388 5.95853 8.9869C5.79449 8.88796 5.6297 8.79023 5.46426 8.6937C5.43455 8.67635 5.42288 8.65584 5.42307 8.62159C5.42427 8.43114 5.42362 8.2406 5.42372 8.05006C5.42372 8.03623 5.42557 8.0224 5.42733 7.9978Z"
-          fill="white"
+          fill="#F0B90B"
         />
         <path
           d="M5.43381 4.49368C5.44862 4.4832 5.45927 4.47447 5.47093 4.46751C5.63331 4.37099 5.79616 4.2753 5.95798 4.17794C5.98585 4.16114 6.00797 4.15975 6.03658 4.17692C6.19859 4.27409 6.36162 4.36941 6.52418 4.46556C6.53622 4.47271 6.54742 4.48125 6.56427 4.49276C6.55103 4.50352 6.54252 4.51243 6.53224 4.51856C6.36486 4.61777 6.19702 4.71634 6.02973 4.81574C6.00742 4.82901 5.98872 4.8292 5.96631 4.81574C5.7993 4.71587 5.63165 4.61703 5.46436 4.51754C5.45501 4.51197 5.44695 4.50408 5.43372 4.49368H5.43381Z"
-          fill="white"
+          fill="#F0B90B"
         />
         <path
           d="M5.00601 7.25245C4.98619 7.2428 4.9736 7.2377 4.96194 7.23083C4.80021 7.13607 4.63884 7.04056 4.47665 6.94664C4.44878 6.93049 4.43573 6.91193 4.43592 6.87861C4.43712 6.6851 4.43647 6.49149 4.43666 6.29798C4.43666 6.28545 4.43888 6.27301 4.44073 6.25269C4.46156 6.26225 4.47822 6.26819 4.49313 6.277C4.63736 6.36211 4.78169 6.44694 4.92509 6.53326C5.02443 6.59312 5.00841 6.56351 5.00906 6.68324C5.01008 6.85383 5.00934 7.02442 5.00915 7.1951C5.00915 7.21152 5.0073 7.22795 5.00582 7.25245H5.00601Z"
-          fill="white"
+          fill="#F0B90B"
         />
         <path
           d="M6.99475 7.25017C6.99271 7.22993 6.99012 7.2162 6.99012 7.20256C6.98984 7.01062 6.99049 6.81878 6.98947 6.62684C6.98929 6.59399 7.00179 6.57459 7.02965 6.55826C7.19009 6.46442 7.3496 6.36911 7.50957 6.27453C7.52152 6.26748 7.53429 6.26182 7.55281 6.25244C7.55595 6.26961 7.55994 6.28131 7.55994 6.293C7.56031 6.4906 7.55994 6.68829 7.56049 6.88588C7.56058 6.91382 7.54901 6.93052 7.52568 6.94417C7.36025 7.0406 7.19527 7.13796 7.03002 7.23485C7.02169 7.23977 7.01215 7.24265 6.99466 7.25017H6.99475Z"
-          fill="white"
+          fill="#F0B90B"
         />
       </g>
       <defs>
-        <clipPath id="clip0_6150_47860">
+        <clipPath id="clip0_6150_47876">
           <rect
             width="5.10638"
             height="6"
