+++ src/components/Icons/Svg/FT/FTIcons.stories.tsx
@@ -4,13 +4,13 @@ import { USDC } from './USDC';
 import { FT1 } from './FT1';
 import { WMAS } from './WMAS';
 import { WETH } from './WETH';
-import { Sepolia } from './chainIcons/Sepolia';
-import { SepoliaBridged } from './chainIcons/SepoliaBridged';
-import { Eth } from './chainIcons/Eth';
-import { EthBridged } from './chainIcons/EthBridged';
+import { Sepolia } from '../ChainIcons/Sepolia';
+import { SepoliaBridged } from '../ChainIcons/SepoliaBridged';
+import { Eth } from '../ChainIcons/Eth';
+import { EthBridged } from '../ChainIcons/EthBridged';
 
-import { BscBridged } from './chainIcons/BscBridged';
-import { Bsc } from './chainIcons/Bsc';
+import { BscBridged } from '../ChainIcons/BscBridged';
+import { Bsc } from '../ChainIcons/Bsc';
 import { createBridgedFt, createNativeFt, getAssetIcons } from './tokenIcons';
 import { bsc, mainnet, sepolia } from 'viem/chains';
 export default { title: 'Icons/FT' };
