+++ src/components/Icons/Svg/FT/index.ts
@@ -5,7 +5,5 @@ export * from './WETH';
 export * from './FT1';
 export * from './WMAS';
 
-export * from './chainIcons';
-
 export * from './tokenIcons';
 export * from './svgInterface';
