+++ src/components/Icons/Svg/index.ts
@@ -1,2 +1,3 @@
 export * from './Massa';
 export * from './FT';
+export * from './ChainIcons';
